pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartdesk"
version = "0.1.0"
description = "SmartDesk — a Python productivity assistant"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Bhavesh Shah" }]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "requests>=2.28",
    "python-dotenv>=1.0"
]
